move birthday command in own package
